refactor: handlers for autorization
